# [Unit Tests] Add Text Layout Performance Tests