--- a/AsyncDisplayKitTests/ASTextNodePerformanceTests.m
+++ b/AsyncDisplayKitTests/ASTextNodePerformanceTests.m
@@ -55,7 +55,7 @@ static NSString *const kTestCaseUIKitWithReusedContext = @"UIKitReusedContext";
   ctx.results[kTestCaseASDK].userInfo[@"size"] = NSStringFromCGSize(asdkSize);
   
   ASXCTAssertEqualSizes(uiKitSize, asdkSize);
-  ASXCTAssertRelativePerformanceInRange(ctx, kTestCaseASDK, 0.25, 0.5);
+  ASXCTAssertRelativePerformanceInRange(ctx, kTestCaseASDK, 0.2, 0.5);
 }
 
 - (void)testPerformance_OneParagraphLatinWithTruncation