This will hopefully help us monitor and improve our text layout performance. There is certainly a ton of room for improvement. 

The tests are currently single-threaded. In the future we could beef em up to be multithreaded. 

Ready for review @maicki @appleguy @garrettmoon @levi 

Here are the current results on my machine:

```
Test Suite 'Selected tests' started at 2016-08-28 15:26:24.649
Test Suite 'ASTextNodePerformanceTests' started at 2016-08-28 15:26:24.650
Test Case '-[ASTextNodePerformanceTests testPerformance_OneParagraphLatinWithTruncation]' started.
Results:
	<UIKit               : time-per-1000=0.06 rel-perf=1.00 user-info={     size = "{348, 134}"; }>
	<ASDK                : time-per-1000=0.47 rel-perf=0.13 user-info={     size = "{344, 134}"; }>
Test Case '-[ASTextNodePerformanceTests testPerformance_OneParagraphLatinWithTruncation]' passed (5.616 seconds).
Test Case '-[ASTextNodePerformanceTests testPerformance_TwoParagraphLatinNoTruncation]' started.
Results:
	<UIKit               : time-per-1000=0.13 rel-perf=1.00 user-info={     size = "{355, 468}"; }>
	<ASDK                : time-per-1000=0.48 rel-perf=0.27 user-info={     size = "{355, 468}"; }>
Test Case '-[ASTextNodePerformanceTests testPerformance_TwoParagraphLatinNoTruncation]' passed (6.506 seconds).
Test Case '-[ASTextNodePerformanceTests testThatNotUsingAStringDrawingContextHasSimilarPerformanceToHavingOne]' started.
Results:
	<UIKitFreshContext   : time-per-1000=0.06 rel-perf=1.02 user-info={     size = "{347.703125, 133.65625}"; }>
	<UIKitNoContext      : time-per-1000=0.06 rel-perf=1.00 user-info={     size = "{347.703125, 133.65625}"; }>
	<UIKitReusedContext  : time-per-1000=0.06 rel-perf=1.03 user-info={     size = "{347.703125, 133.65625}"; }>
Test Case '-[ASTextNodePerformanceTests testThatNotUsingAStringDrawingContextHasSimilarPerformanceToHavingOne]' passed (1.843 seconds).
Test Case '-[ASTextNodePerformanceTests testThatUIKitPrivateLayoutCachingIsAwesome]' started.
Results:
	<UIKit               : time-per-1000=0.06 rel-perf=1.00 user-info={     size = "{347.703125, 133.65625}"; }>
	<UIKitPrivateCaching : time-per-1000=0.03 rel-perf=2.39 user-info={     size = "{347.703125, 133.65625}"; }>
Test Case '-[ASTextNodePerformanceTests testThatUIKitPrivateLayoutCachingIsAwesome]' passed (0.895 seconds).
```