@Adlai-Holler 2016-08-28
Hmm my VPN isn't accessible so I can't tell where Jenkins is failing. I assume the performance variation is just broader and I need to soften some of the expectations, but I wish I knew which…