@maicki 2016-08-28
@Adlai-Holler Do you know `dispatch_benchmark`?

`curl "http://opensource.apple.com/source/libdispatch/libdispatch-84.5/man/dispatch_benchmark.3?txt" | /usr/bin/groffer --tty -T utf8`