@maicki 2016-08-28
👍

> On Aug 28, 2016, at 4:29 PM, Adlai Holler <notifications@github.com> wrote:
> 
> Yes but I can't use it here because it just isn't flexible enough for e.g. controlling which sections are benchmarked, controlling thread count, and eventually returning values. Better to roll our own I believe.
> 
> Sent from my iPhone
> 
> > On Aug 28, 2016, at 4:18 PM, Michael Schneider <notifications@github.com> wrote:
> > 
> > @Adlai-Holler Do you know dispatch_benchmark?
> > 
> > curl "http://opensource.apple.com/source/libdispatch/libdispatch-84.5/man/dispatch_benchmark.3?txt" | /usr/bin/groffer --tty -T utf8
> > 
> > —
> > You are receiving this because you were mentioned.
> > Reply to this email directly, view it on GitHub, or mute the thread.
> > 
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub, or mute the thread.
>